Move bootstrap config to a separate file.
